+++ lib/cache.js
@@ -1,7 +1,7 @@
 var SortedArrayMap = require("collections/sorted-array-map"),
     Map = require("collections/map"),
     Promise = require("bluebird");
-// TypeError = require("./Errors.js").TypeError;
+// TypeError = require("./errors.js").TypeError;
 
 //////////////////////////////////////////// Time Ordered Array ////////////////////////////////////////////
