+++ lib/utils.js
@@ -1,6 +1,6 @@
 var url = require('url'),
     parseUrl = require('url').parse,
-    InvalidStateError = require('./Errors').InvalidStateError;
+    InvalidStateError = require('./errors').InvalidStateError;
 
 
 var redefine = function (obj, prop, value) {
